Refactored _mkl_interface.py into a subpackage
